+++ src/modules/cron/cron.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { UsersModule } from '../users/users.module';
+import { CronService } from './cron.service';
+
+@Module({
+  imports: [UsersModule],
+  providers: [CronService],
+  exports: [CronService]
+})
+export class CronModule {}